requirements.txt
aiocsv==1.2.3
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
astroid==2.12.12
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
bacpypes==0.18.6
certifi==2022.9.24
charset-normalizer==2.1.1
ciso8601==2.2.0
click==8.1.3
dill==0.3.6
fastapi==0.87.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
influxdb-client==1.34.0
isort==5.10.1
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
mypy==0.991
mypy-extensions==0.4.3
platformdirs==2.5.4
ply==3.11
pycodestyle==2.10.0
pycryptodomex==3.15.0
pydantic==1.10.2
pylint==2.15.6
pymodbus==3.0.2
pyserial==3.5
pyserial-asyncio==0.6
pysnmp-pyasn1==1.1.2
pysnmp-pysmi==1.1.10
pysnmplib==5.0.20
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
reactivex==4.0.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tomlconfig @ git+https://github.com/JurajMarcin/tomlconfig.git@1.1.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
yarl==1.8.1
